+++ tests/integration/Contrib.FromNodePackageJson.Builders.ToolBuilder.node.test.js
@@ -22,18 +22,17 @@ const assert = require('node:assert')
 const { suite, test } = require('mocha')
 
 const {
+  Contrib,
   Models,
-  Factories,
-  Builders: { FromNodePackageJson: { ToolBuilder } }
 } = require('../../')
 
-suite('integration: Builders.FromNodePackageJson.ToolBuilder', () => {
+suite('integration: Contrib.FromNodePackageJson.Builders.ToolBuilder', () => {
   const salt = Math.random()
 
-  const extRefFactory = new Factories.FromNodePackageJson.ExternalReferenceFactory()
+  const extRefFactory = new Contrib.FromNodePackageJson.Factories.ExternalReferenceFactory()
   extRefFactory.makeExternalReferences = () => [`FAKE REFERENCES ${salt}`]
 
-  const sut = new ToolBuilder(extRefFactory)
+  const sut = new Contrib.FromNodePackageJson.Builders.ToolBuilder(extRefFactory)
 
   const data = {
     name: '@foo/bar',